backend/export-sidecar/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-docx==1.1.2
latex2word==1.1
latex2mathml==3.77.0
lxml==5.3.2
Pillow==12.1.0
httpx==0.28.1
nacos-sdk-python==1.0.0
mysql-connector-python==9.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
